% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_utils.R
\name{dist.cohen.d}
\alias{dist.cohen.d}
\title{Calculate Cohen's d for Distance Matrix}
\usage{
dist.cohen.d(dm, f)
}
\arguments{
\item{dm}{Distance matrix.}

\item{f}{Factor variable with two levels.}
}
\value{
Cohen's d value if factor has exactly two levels; NULL otherwise.
}
\description{
This function calculates the Cohen's d statistic for a distance matrix,
assuming two levels in the factor variable.
}
\examples{
if (requireNamespace("phyloseq", quietly = TRUE)) {
  data("enterotype", package = "phyloseq")
  ent <- phyloseq::subset_samples(enterotype, !is.na(Enterotype))
  dm <- phyloseq::distance(ent, method = "bray")
  meta <- data.frame(phyloseq::sample_data(ent))
  f <- factor(meta$Enterotype)

  keep <- f \%in\% c("1", "2")
  dm_12 <- as.dist(as.matrix(dm)[keep, keep])
  f_12 <- droplevels(f[keep])

  dist.cohen.d(dm_12, f_12)
}

}
