% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_utils.R
\name{dist.group.sigma2}
\alias{dist.group.sigma2}
\title{Calculate Group-wise Sigma Squared}
\usage{
dist.group.sigma2(dm, f)
}
\arguments{
\item{dm}{Distance matrix.}

\item{f}{Factor variable for group definition.}
}
\value{
A named numeric vector of group-wise sigma squared values.
}
\description{
This function calculates the sigma squared statistic for each group defined
by a factor variable in a given distance matrix.
}
\examples{
if (requireNamespace("phyloseq", quietly = TRUE)) {
  data("enterotype", package = "phyloseq")
  ent <- phyloseq::subset_samples(enterotype, !is.na(Enterotype))
  dm <- phyloseq::distance(ent, method = "bray")
  meta <- data.frame(phyloseq::sample_data(ent))
  f <- factor(meta$Enterotype)

  dist.group.sigma2(dm, f)
}
}
