% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_utils.R
\name{dist.sigma2}
\alias{dist.sigma2}
\title{Calculate Sigma Squared for Distance Matrix}
\usage{
dist.sigma2(dm)
}
\arguments{
\item{dm}{Distance matrix.}
}
\value{
Sigma squared value.
}
\description{
This function computes the sigma squared statistic for a given distance matrix.
}
\examples{

if (requireNamespace("phyloseq", quietly = TRUE)) {
  data("enterotype", package = "phyloseq")
  ent <- phyloseq::subset_samples(enterotype, !is.na(Enterotype))
  dm <- phyloseq::distance(ent, method = "bray")

  dist.sigma2(dm)
}

}
