% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_utils.R
\name{dist.ss2}
\alias{dist.ss2}
\title{Calculate Sum of Squares for Distance Matrix with Factor}
\usage{
dist.ss2(dm2, f)
}
\arguments{
\item{dm2}{Square distance matrix.}

\item{f}{Factor variable for weighting.}
}
\value{
Sum of squares matrix.
}
\description{
This function calculates the sum of squares for a given distance matrix,
weighted by a factor variable.
}
\examples{

if (requireNamespace("phyloseq", quietly = TRUE)) {
  data("enterotype", package = "phyloseq")
  ent <- phyloseq::subset_samples(enterotype, !is.na(Enterotype))
  dm <- phyloseq::distance(ent, method = "bray")
  meta <- data.frame(phyloseq::sample_data(ent))
  f <- factor(meta$Enterotype)

  dist.ss2(as.matrix(dm)^2, f)
}

}
