% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{generic.distance.permutation.test}
\alias{generic.distance.permutation.test}
\title{Conduct a Generic Distance-Based Permutation Test}
\usage{
generic.distance.permutation.test(
  test.statistic,
  dm,
  f,
  nrep = 999,
  strata = NULL
)
}
\arguments{
\item{test.statistic}{A function that takes \code{dm} and \code{f} and
returns the observed test statistic.}

\item{dm}{A distance matrix (any arbitrary distance or dissimilarity metric).}

\item{f}{A factor variable representing the groups.}

\item{nrep}{The number of permutations to conduct. Default is 999.}

\item{strata}{A factor variable representing permutation strata. If supplied,
permutations are restricted within strata. Default is \code{NULL}.}
}
\value{
A list containing:
\itemize{
  \item \code{p.value}: The p-value of the test
  \item \code{statistic}: The observed test statistic
  \item \code{nrep}: The number of permutations performed
}
}
\description{
This function performs a distance-based permutation test using an arbitrary
test statistic. It is useful when a statistic such as \code{\link{WdS}} or
\code{\link{Tw2}} has been defined separately and permutation-based
significance testing is needed for multivariate distance data.
}
\examples{
if (requireNamespace("phyloseq", quietly = TRUE)) {
  data("enterotype", package = "phyloseq")
  ent <- phyloseq::subset_samples(enterotype, !is.na(Enterotype))
  dm <- phyloseq::distance(ent, method = "bray")
  meta <- data.frame(phyloseq::sample_data(ent))
  f <- factor(meta$Enterotype)

  generic.distance.permutation.test(WdS, dm = dm, f = f, nrep = 9)
}
}
\references{
Hamidi B, Wallace K, Vasu C, Alekseyenko AV. The
\eqn{\mathnormal{W}_d^*}-test: robust distance-based multivariate analysis
of variance. \emph{Microbiome}. 2019;7:51.
}
\seealso{
\code{\link{Tw2.test}}, \code{\link{WdS.test}}
}
