% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_utils.R
\name{is.dist}
\alias{is.dist}
\title{Test if an object is of class 'dist'}
\usage{
is.dist(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical indicating whether the object is a distance matrix.
}
\description{
This function checks whether the given object belongs to class 'dist'.
}
\examples{
\dontrun{
is.dist(as.dist(matrix(1:4, nrow = 2)))
is.dist(matrix(1:4, nrow = 2))
}
}
