% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{Tw2.test}
\alias{Tw2.test}
\title{Conduct a \eqn{T_w^2} Distance-Based Permutation Test}
\usage{
Tw2.test(dm, f, nrep = 999)
}
\arguments{
\item{dm}{A distance matrix (any arbitrary distance or dissimilarity metric).}

\item{f}{A factor variable representing the groups.}

\item{nrep}{The number of permutations to conduct. Default is 999.}
}
\value{
A list containing:
\itemize{
  \item \code{p.value}: The p-value of the test
  \item \code{statistic}: The observed test statistic
  \item \code{nrep}: The number of permutations performed
}
}
\description{
This function is a specialized version of
\code{\link{generic.distance.permutation.test}} that uses the
\code{\link{Tw2}} statistic for a two-group comparison.
}
\examples{
if (requireNamespace("phyloseq", quietly = TRUE)) {
  data("enterotype", package = "phyloseq")
  ent <- phyloseq::subset_samples(enterotype, !is.na(Enterotype))
  dm <- phyloseq::distance(ent, method = "bray")
  meta <- data.frame(phyloseq::sample_data(ent))
  f <- factor(meta$Enterotype)

  keep <- f \%in\% c("1", "2")
  dm_12 <- as.dist(as.matrix(dm)[keep, keep])
  f_12 <- droplevels(f[keep])

  Tw2.test(dm_12, f_12, nrep = 9)
}
}
\seealso{
\code{\link{Tw2}}, \code{\link{WdS.test}},
  \code{\link{generic.distance.permutation.test}}
}
