% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_utils.R
\name{a.dist}
\alias{a.dist}
\title{Covariate Adjusted Principal Coordinates Analysis}
\usage{
a.dist(formula, data = parent.frame())
}
\arguments{
\item{formula}{A typical formula such as Y~ A, but here Y is a dissimilarity distance.
The formula has the same requirements as in adonis function of the vegan package.}

\item{data}{A dataset with the rownames the same as the rownames in distance.
This dataset should include both the confounding covariate and the primary covariate.
If not provided, the parent data.frame will be used.}
}
\value{
Returns a distance matrix of class `"dist"` representing the Euclidean distances
}
\description{
This function takes in a formula and distance matrix, adjusts for covariates,
performs PCoA, and returns the resuting corrected matrix
}
\details{
The 'a.dist' function first evaluates the left-hand side (LHS) of the formula
         within the data frame. Then, it constructs a model matrix from the right-hand side
         (RHS) of the formula. After performing necessary matrix operations and eigen
         decomposition, it calculates the Euclidean distances
}
\examples{
\dontrun{
data(iris)
formula <- Species ~ Sepal.Length + Sepal.Width + Petal.Length + Petal.Width
dist_matrix <- a.dist(formula, iris)
print(dist_matrix)}
}
