% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{aWdS.test}
\alias{aWdS.test}
\title{Conducts an adjusted WdS distance-based multivariate Welch ANOVA using optional confounder elimination}
\usage{
aWdS.test(
  dm,
  f,
  nrep = 999,
  strata = NULL,
  formula = NULL,
  formula_data = parent.frame()
)
}
\arguments{
\item{dm}{A distance matrix (any arbitrary distance or dissimilarity metric).}

\item{f}{A factor variable representing the groups.}

\item{nrep}{The number of permutations to conduct. Default is 999.}

\item{strata}{(optional) A factor variable representing strata. If specified,
the test will perform p-value computations using stratified permutation.
Default is NULL.}

\item{formula}{(optional) A right-hand side ONLY formula to be used with 
'formula_data' for confounder adjustment of WdS statistic, Omega squared (ω²)
 effect size estimate, and p-value. Default is NULL}

\item{formula_data}{(optional) A data frame to be used in conjunction with 
'formula' for confounder adjustment. Default is parent.frame()}
}
\value{
A list containing:
\itemize{
  \item \code{method}: The name of the method used.
  \item \code{data.name}: A string containing the name of input data.
  \item \code{statistic}: The observed WdS test statistic.
  \item \code{estimate}: The Omega squared (ω²) effect size estimate.
  \item \code{nrep}: The number of permutations performed.
  \item \code{p.value}: The p-value of the test.
  \item \code{parameter}: a list of 2-4 containing strata (optional), 
  adjustment formula (optional), between degrees of freedom (dfb), and number of permutations performed
  (nrep).
}
}
\description{
'aWdS.test()' defaults to an unadjusted Wds.test() if the optional parameters 'formula' and 'formula_data' are not provided. 
Users may choose to pre-adjust their 'dm' before using this function or choose to use make adjustments within the function.
}
\details{
This function performs the WdS test statistic for k-group comparison using a given distance matrix.
Optionally, it can preprocess the provided data through the \code{a.dist} function using a specified
formula before performing the test, if 'dm', 'formula', and 'formula_data' are provided.  

If optional adjustment parameters are provided, it projects matrices to remove the effect of covariate.
}
\examples{
data(mtcars)

# The outcome could be a single variable or multiple variables (such as multidimensional omics data).  
## This is an example with a single variable:  
dm <- dist(mtcars$mpg, method="euclidean")

## This is an example with multiple variables:  
dm <- dist(mtcars[1:3], method="euclidean") 

f <- factor(mtcars$gear)

# Adjusted example using *'aWdS.test()'*  
## Right-hand side adjustment formula. Note that you may use any data type including factor, character, integer, and numeric.  
formula <- ~ wt + as.factor(am) 
aWdS.test(dm=dm, f=f, formula=formula, formula_data=mtcars)
# Adjusted example using *'WdS.test()'*  
## Create the adjusted distance matrix 'a.dm'  
a.dm <- a.dist(dm=dm, formula=formula, formula_data=mtcars)
# Perform adjusted test with 'a.dm'  
WdS.test(dm=a.dm, f=f)

# Unadjusted example (equivalent to 'WdS.test()')  
aWdS.test(dm=dm, f=f)

}
\seealso{
\code{\link{WdS.test}}, \code{\link{generic.distance.permutation.test}}, \code{\link{Tw2.test}},
         \code{\link{a.dist}}
}
