% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_utils.R
\name{dist.cohen.d}
\alias{dist.cohen.d}
\title{Calculate Cohen's d for Distance Matrix}
\usage{
dist.cohen.d(dm, f)
}
\arguments{
\item{dm}{Distance matrix.}

\item{f}{Factor variable with two levels.}
}
\value{
Cohen's d value if factor has exactly two levels; NULL otherwise.
}
\description{
This function calculates the Cohen's d statistic for a distance matrix,
assuming two levels in the factor variable.
}
\examples{
dm <- as.dist(matrix(runif(100), nrow = 10))
f <- factor(c(rep("A", 5), rep("B", 5)))
dist.cohen.d(dm, f)
}
