% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_utils.R
\name{dist.goodness.of.fit}
\alias{dist.goodness.of.fit}
\title{Calculate Goodness of Fit for Adjusted Distance Matrices}
\usage{
dist.goodness.of.fit(dm, adjusted_dm)
}
\arguments{
\item{dm}{The original/raw distance matrix.}

\item{adjusted_dm}{The adjusted/residual distance matrix.}
}
\value{
Goodness of fit (\eqn{R^2}) coefficient of prediction.
}
\description{
This function computes the coefficient of prediction (\eqn{R^2}) by
comparing total variation in a raw distance matrix with residual variation in
an adjusted distance matrix.
}
\examples{
data(mtcars)
dm <- dist(mtcars[1:3], method = "euclidean")
adjusted_dm <- a.dist(dm, formula = ~ wt, formula_data = mtcars)
dist.goodness.of.fit(dm, adjusted_dm)

}
