% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_utils.R
\name{dist.goodness.of.fit}
\alias{dist.goodness.of.fit}
\title{Calculate Goodness-of-Fit for Residual Distance Matrices}
\usage{
dist.goodness.of.fit(dm, dm_residual)
}
\arguments{
\item{dm}{The original/raw distance matrix.}

\item{dm_residual}{The residual distance matrix.}
}
\value{
Goodness-of-fit coefficient of determination (\eqn{R^2})
}
\description{
This function computes the coefficient of determination (\eqn{R^2}) by
comparing total variation in a raw distance matrix with residual variation in
a residual distance matrix.
}
\examples{
data(mtcars)
dm <- dist(mtcars[1:3], method = "euclidean")
dm_residual <- a.dist(dm, formula = ~ wt, formula_data = mtcars)
dist.goodness.of.fit(dm, dm_residual)

}
