% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_utils.R
\name{dist.sigma2}
\alias{dist.sigma2}
\title{Calculate Sigma Squared for Distance Matrix}
\usage{
dist.sigma2(dm)
}
\arguments{
\item{dm}{Distance matrix.}
}
\value{
Sigma squared value.
}
\description{
This function computes the sigma squared statistic for a given distance matrix.
}
\examples{
\dontrun{
dm <- as.dist(matrix(runif(100), nrow = 10))
dist.sigma2(dm)
}
}
