% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_utils.R
\name{dist.ss2}
\alias{dist.ss2}
\title{Calculate Sum of Squares for Distance Matrix with Factor}
\usage{
dist.ss2(dm2, f)
}
\arguments{
\item{dm2}{Square distance matrix.}

\item{f}{Factor variable for weighting.}
}
\value{
Sum of squares matrix.
}
\description{
This function calculates the sum of squares for a given distance matrix,
weighted by a factor variable.
}
\examples{

dm <- matrix(runif(100), nrow = 10)
f <- factor(c(rep("A", 5), rep("B", 5)))
dist.ss2(dm, f)

}
