% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{generic.distance.permutation.test}
\alias{generic.distance.permutation.test}
\title{Conducts a generic distance-based permutation test for k-group differences}
\usage{
generic.distance.permutation.test(
  test.statistic,
  dm,
  f,
  nrep = 999,
  strata = NULL
)
}
\arguments{
\item{test.statistic}{A function to calculate the test statistic.}

\item{dm}{A distance metric (any arbitrary distance or dissimilarity metric).}

\item{f}{A factor variable representing the groups.}

\item{nrep}{The number of permutations to conduct. Default is 999.}

\item{strata}{An optional stratifying variable for restricted permutation.}
}
\value{
A list containing:
\itemize{
  \item \code{p.value}: The p-value of the test
  \item \code{statistic}: The observed test statistic
  \item \code{nrep}: The number of permutations performed
}
}
\description{
This function performs a distance-based permutation test using an arbitrary test statistic.
It is built on the principle of Welch's ANOVA and extends it to handle multivariate distance data.
It is particularly useful for analyzing microbiome data.
}
\examples{
# TODO: Add examples
}
\references{
Hamidi, Bashir, et al. "$ W_ {d}^{*} $-test: robust distance-based multivariate analysis of variance." Microbiome 7.1 (2019): 1-9.
}
\seealso{
\code{\link{Tw2.test}}, \code{\link{WdS.test}}
}
