% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc_utils.R
\name{Tw2.posthoc.1vsAll.tests}
\alias{Tw2.posthoc.1vsAll.tests}
\title{Perform 1-vs-All posthoc tests for Tw2 statistic}
\usage{
Tw2.posthoc.1vsAll.tests(dm, f, nrep = 999, strata = NULL)
}
\arguments{
\item{dm}{A distance matrix.}

\item{f}{A factor variable.}

\item{nrep}{Number of permutations (default: 999).}

\item{strata}{A factor variable for stratified permutation (default: NULL).}
}
\value{
A matrix containing 1-vs-All posthoc test results.
}
\description{
This function performs 1-vs-All posthoc tests for the Tw2 statistic using a permutation approach.
}
\examples{
data <- ...  # TODO add data here
Tw2.posthoc.1vsAll.tests(data$dm, data$f)

}
