% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc_utils.R
\name{Tw2.posthoc.tests}
\alias{Tw2.posthoc.tests}
\title{Perform posthoc tests for Tw2 statistic}
\usage{
Tw2.posthoc.tests(dm, f, nrep = 999, strata = NULL)
}
\arguments{
\item{dm}{A distance matrix.}

\item{f}{A factor variable.}

\item{nrep}{Number of permutations (default: 999).}

\item{strata}{A factor variable for stratified permutation (default: NULL).}
}
\value{
A matrix containing posthoc test results.
}
\description{
This function performs posthoc tests for the Tw2 statistic using a permutation approach.
}
\examples{
# Example usage
data <- ...  # TODO add data here
Tw2.posthoc.tests(data$dm, data$f)

}
