% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc_utils.R
\name{Tw2.posthoc.tests}
\alias{Tw2.posthoc.tests}
\title{Perform posthoc tests for Tw2 statistic}
\usage{
Tw2.posthoc.tests(dm, f, nrep = 999)
}
\arguments{
\item{dm}{A distance matrix representing the pairwise distances between observations.}

\item{f}{A factor variable indicating the group membership of each observation.}

\item{nrep}{Number of permutations to perform (default: 999).}
}
\value{
A matrix containing posthoc test results, with columns for level combinations,
        sample sizes, p-values, Tw2 statistics, and number of permutations.
}
\description{
This function performs posthoc tests for the Tw2 statistic using a permutation approach.
The Tw2 statistic is an extension of the Welch ANOVA test for multivariate distances.
It is especially useful for analyzing microbiome data.
}
\examples{
\dontrun{
# Example usage
data <- ... # TODO add data here
result <- Tw2.posthoc.tests(data$dm, data$f)
print(result)
}

}
\seealso{
\url{https://github.com/alekseyenko/WdStar}
}
