% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{Tw2.test}
\alias{Tw2.test}
\title{Conducts a Tw2 distance-based permutation test for k-group differences}
\usage{
Tw2.test(dm, f, nrep = 999)
}
\arguments{
\item{dm}{A distance metric (any arbitrary distance or dissimilarity metric)}

\item{f}{A factor variable representing the groups.}

\item{nrep}{The number of permutations to conduct. Default is 999.}
}
\value{
A list containing:
\itemize{
  \item \code{p.value}: The p-value of the test
  \item \code{statistic}: The observed test statistic
  \item \code{nrep}: The number of permutations performed
}
}
\description{
This function is a specialized version of the \code{\link{generic.distance.permutation.test}},
specifically designed to use the Tw2 test statistic for k-group comparison.
}
\examples{
# TODO: Add examples
}
\seealso{
\code{\link{generic.distance.permutation.test}}, \code{\link{WdS.test}}
}
