% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{WdS}
\alias{WdS}
\title{This method computes Wd* statistic for distance matrix dm and factor f}
\usage{
WdS(dm, f)
}
\arguments{
\item{dm}{Distance matrix}

\item{f}{Factor}
}
\description{
This method computes Wd* statistic for distance matrix dm and factor f
}
