% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{WdS.test}
\alias{WdS.test}
\title{Conducts a WdS Distance-Based Permutation Test for k-Group Differences with Optional aPCoA Preprocessing}
\usage{
WdS.test(dm, f, nrep = 999, strata = NULL, data = NULL, formula = NULL)
}
\arguments{
\item{dm}{A distance matrix (any arbitrary distance or dissimilarity metric).}

\item{f}{A factor variable representing the groups.}

\item{nrep}{The number of permutations to conduct. Default is 999.}

\item{strata}{An optional stratifying variable for restricted permutation.}

\item{data}{(optional) Data frame to be used in conjunction with 'formula' and 'dm' for aPCoA preprocessing.}

\item{formula}{(optional) A formula to be used with 'data' for aPCoA preprocessing.}
}
\value{
A list containing:
\itemize{
  \item \code{p.value}: The p-value of the test.
  \item \code{statistic}: The observed test statistic.
  \item \code{nrep}: The number of permutations performed.
}
}
\description{
This function performs the WdS test statistic for k-group comparison using a given distance matrix.
Optionally, it can preprocess the provided data through the \code{aPCoA.dist} function using a specified
formula before performing the test, if 'data', 'dm', and 'formula' are provided.
}
\examples{
# Example with provided distance matrix
\dontrun{
dm <- as.dist(matrix(runif(100), nrow = 10))
f <- factor(c(rep("A", 5), rep("B", 5)))
WdS.test(dm, f)

# Example with optional aPCoA preprocessing
data(iris)
dm <- as.dist(matrix(runif(100), nrow = 10))
formula <- Species ~ Sepal.Length + Sepal.Width + Petal.Length + Petal.Width
WdS.test(dm = dm, f = iris$Species, data = iris, formula = formula)
}

}
\seealso{
\code{\link{generic.distance.permutation.test}}, \code{\link{Tw2.test}},
         \code{\link{aPCoA.dist}}
}
