% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{WdS.test}
\alias{WdS.test}
\title{Conducts a WdS Distance-Based Permutation Test for k-Group Differences with Optional confounder elimination}
\usage{
WdS.test(dm, f, nrep = 999, data = NULL, formula = NULL)
}
\arguments{
\item{dm}{A distance matrix (any arbitrary distance or dissimilarity metric).}

\item{f}{A factor variable representing the groups.}

\item{nrep}{The number of permutations to conduct. Default is 999.}

\item{data}{(optional) Data frame to be used in conjunction with 'formula' and 'dm' for confounder elimination.}

\item{formula}{(optional) A formula to be used with 'data' for confounder elimination}
}
\value{
A list containing:
\itemize{
  \item \code{p.value}: The p-value of the test.
  \item \code{statistic}: The observed test statistic.
  \item \code{nrep}: The number of permutations performed.
}
}
\description{
This function performs the WdS test statistic for k-group comparison using a given distance matrix.
Optionally, it can preprocess the provided data through the \code{a.dist} function using a specified
formula before performing the test, if 'data', 'dm', and 'formula' are provided.
}
\examples{
# Example with provided distance matrix
\dontrun{
dm <- as.dist(matrix(runif(100), nrow = 10))
f <- factor(c(rep("A", 5), rep("B", 5)))
WdS.test(dm, f)

# Example with optional confounder elimination with a.dist()
data(iris)
formula <- Species ~ Sepal.Length + Sepal.Width + Petal.Length + Petal.Width

distance_matrix = dist(iris[2:4], method="euclidean") # Numerical columns only

WdS.test(dm = distance_matrix, f = iris$Species)}
}
\seealso{
\code{\link{generic.distance.permutation.test}}, \code{\link{Tw2.test}},
         \code{\link{a.dist}}
}
