% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{WdS.test}
\alias{WdS.test}
\title{Conducts a WdS distance-based permutation test for k-group differences}
\usage{
WdS.test(dm, f, nrep = 999, strata = NULL)
}
\arguments{
\item{dm}{A distance or dissimilarity matrix.}

\item{f}{A factor variable representing the groups.}

\item{nrep}{The number of permutations to conduct. Default is 999.}

\item{strata}{An optional stratifying variable for restricted permutation.}
}
\value{
A list containing:
\itemize{
  \item \code{p.value}: The p-value of the test
  \item \code{statistic}: The observed test statistic
  \item \code{nrep}: The number of permutations performed
}
}
\description{
This function is a specialized version of the \code{\link{generic.distance.permutation.test}},
specifically designed to use the WdS test statistic for k-group comparison.
}
\examples{
# TODO: Add examples
}
\seealso{
\code{\link{generic.distance.permutation.test}}, \code{\link{Tw2.test}}
}
