% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{WdS.test}
\alias{WdS.test}
\title{Conducts the original WdS Distance-Based Permutation Test for k-Group differences}
\usage{
WdS.test(dm, f, nrep = 999, strata = NULL)
}
\arguments{
\item{dm}{A distance matrix (any arbitrary distance or dissimilarity metric).}

\item{f}{A factor variable representing the groups.}

\item{nrep}{The number of permutations to conduct. Default is 999.}

\item{strata}{(optional) A factor variable to be used for stratified permutation.}
}
\value{
A list containing:
\itemize{
  \item \code{p.value}: The p-value of the test.
  \item \code{statistic}: The observed test statistic.
  \item \code{nrep}: The number of permutations performed.
}
}
\description{
This function performs the WdS test statistic for k-group comparison using a given distance matrix.
}
\examples{
# Example with provided distance matrix
\dontrun{
dm <- as.dist(matrix(runif(100), nrow = 10))
f <- factor(c(rep("A", 5), rep("B", 5)))
WdS.test(dm, f)
}
}
\seealso{
\code{\link{generic.distance.permutation.test}}, \code{\link{Tw2.test}}
}
