% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{WdS.test}
\alias{WdS.test}
\title{Conducts the original WdS distance-based multivariate Welch ANOVA}
\usage{
WdS.test(dm, f, nrep = 999, strata = NULL)
}
\arguments{
\item{dm}{A distance matrix (any arbitrary distance or dissimilarity metric).}

\item{f}{A factor variable representing the groups.}

\item{nrep}{The number of permutations to conduct. Default is 999.}

\item{strata}{(optional) A factor variable representing strata. If specified,
the test will perform p-value computations using stratified permutation.
Default is NULL.}
}
\value{
A list containing:
\itemize{
  \item \code{method}: The name of the method used.
  \item \code{data.name}: A string containing the name of input data.
  \item \code{statistic}: The observed WdS test statistic.
  \item \code{estimate}: The Omega squared (ω²) effect size estimate.
  \item \code{nrep}: The number of permutations performed.
  \item \code{p.value}: The p-value of the test.
  \item \code{parameter}: a list of 2 or 3 containing Strata (optional),
  between degrees of freedom (dfb), and number of permutations performed
  (nrep).
}
}
\description{
This function performs the WdS test statistic for k-group differences comparison using a given distance matrix.
}
\examples{
data(mtcars)

# The outcome could be a single variable or multiple variables (such as multidimensional omics data).
## This is an example with a single variable:
dm <- dist(mtcars$mpg, method="euclidean")

## This is an example with multiple variables:
dm <- dist(mtcars[1:3], method="euclidean")

f <- factor(mtcars$gear)

# Multivariate test example
WdS.test(dm=dm, f=f)

# Stratified example
strata <- factor(mtcars$vs)
WdS.test(dm=dm, f=f, strata=strata)

# Adjusted example (equivalent to 'aWdS.test()' without the need to pass adjustment variable)
## Right-hand side adjustment formula. Note that you may use any data type including factor, character, integer, and numeric.
formula <- ~ wt + as.factor(am)
## Create the adjusted distance matrix 'a.dm'
a.dm <- a.dist(dm=dm, formula=formula, formula_data=mtcars)
## Perform adjusted test with 'a.dm'
WdS.test(dm=a.dm, f=f)

}
\seealso{
\code{\link{aWdS.test}}, \code{\link{generic.distance.permutation.test}}, \code{\link{Tw2.test}}
}
