% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{WdS.test}
\alias{WdS.test}
\title{Conducts distance-based multivariate Welch ANOVA}
\usage{
WdS.test(
  dm,
  f,
  nrep = 999,
  strata = NULL,
  formula = NULL,
  formula_data = parent.frame(),
  ...
)
}
\arguments{
\item{dm}{A distance matrix (any arbitrary distance or dissimilarity metric).}

\item{f}{A factor variable representing the groups.}

\item{nrep}{(optional) The number of permutations to conduct. Default is 999.}

\item{strata}{(optional) A factor variable representing strata. If specified,
the test will perform p-value computations using stratified permutation.
Default is NULL.}

\item{formula}{(optional) A right-hand side ONLY formula to be used with
'formula_data' for confounder adjustment/elimination from `dm`. Results in
adjusted WdS statistic, omega-squared (\eqn{\omega^2}{\omega^2}) effect
size estimate, model goodness-of-fit coefficient of determination
(\eqn{R^2}), adjustment covariate goodness-of-fit coefficient of
determination (\eqn{R^2}), and p-value. Note that you may use any data type
including factor, character, integer, and numeric. Default is NULL}

\item{formula_data}{(optional) An environment, data frame, list, or object
coercible to a data frame, such as \code{phyloseq::sample_data()}, to be
used in conjunction with 'formula' for confounder adjustment. Default is
parent.frame()}

\item{...}{Additional arguments passed to \code{\link{a.dist}()}, such as
\code{tol} (tolerance for eigenvalues; default \code{1e-8}). Only used
when \code{formula} is provided.}
}
\value{
A list containing:
\itemize{
  \item \code{method}: name of the method used.
  \item \code{data.name}: a string describing the input data.
  \item \code{statistic}: observed WdS test statistic.
  \item \code{estimate}: effect size estimator of variance, omega-squared (\eqn{\omega^2}{\omega^2}).
  \item \code{model.goodness.of.fit}: model goodness-of-fit coefficient of
  determination (\eqn{R^2}) from the grouping factor \code{f}, computed with
  \code{\link{dist.goodness.of.fit}} as
  \eqn{1 - SS_{residual}/SS_{total}}.
  \item \code{goodness.of.fit}: adjustment covariate goodness-of-fit
  coefficient of determination (\eqn{R^2}) for adjusted tests;
  \code{NULL} for unadjusted tests.
  \item \code{p.value}: The p-value of the test.
  \item \code{parameter}: A list of 2-4 containing:
         \itemize{
         \item \code{strata}: (optional) strata variable used to perform restricted permutations.
         \item \code{formula}: (optional) formula for residual projection to perform matrix adjustment.
         \item \code{dfb}: between degrees of freedom.
         \item \code{nrep}: number of permutations performed
         }
}
}
\description{
This function performs the \eqn{\mathnormal{W}_d^*} test statistic for
k-group comparison using a given distance matrix.
}
\details{
This method uses permutation testing to establish the significance by computing
\eqn{\mathnormal{W}_d^*} statistic on \eqn{\mathnormal{m}} permutations of the original data
and estimate the significance as the fraction of times the permuted statistic
is greater than or equal to \eqn{\mathnormal{W}_d}.

Our approach can accommodate multi-level factors, stratification (via
restricted permutations), multiple post-hoc testing scenarios, and covariate
adjustment/ elimination (via projection of residuals).

If optional covariate adjustment parameters (\code{`formula`}, and
\code{`formula_data`}) are provided as input, they are passed to the \code{WdStar::a.dist()}
function to project residual matrices and remove the effect of specified
covariate(s) from distance matrix \code{dm} before performing the test.
Users may choose to make adjustments within the function or use \code{WdStar::a.dist()}
outside the function and then pass on the adjusted distance matrix to \code{WdStar::WdS.test()}.
}
\examples{
# The following is a simple example using the mtcars dataset to assess the
# effect of gears on mpg, cyl, and disp (first three variables of the dataset):
data(mtcars)

# The outcome could be a single variable or multiple variables (such as
# multidimensional omics data).

## This is an example of outcome with a single variable:
dm <- dist(mtcars$mpg, method="euclidean")

## This is an example of outcome with multiple variables:
dm <- dist(mtcars[1:3], method="euclidean")

# Grouping/independent variable. You could use multiple variables here too.
f <- factor(mtcars$gear)

# Basic multivariate test example ###########
#############################################
WdS.test(dm=dm, f=f)

# Stratified example ########################
#############################################
strata <- factor(mtcars$vs)
WdS.test(dm=dm, f=f, strata=strata)

# Covariate adjustment/elimination examples #
#############################################
## Right-hand side adjustment formula to specify adjustment covariates.
formula <- ~ wt + as.factor(am)

## Adjustment example 1: pass unadjusted `dm` and formula to WdS.test()
WdS.test(dm=dm, f=f, formula=formula, formula_data=mtcars) ## Perform adjusted test

## Adjustment example 2: Create the adjusted distance matrix `a.dm` outside
## the function
a.dm <- a.dist(dm=dm, formula=formula, formula_data=mtcars)
WdS.test(dm=a.dm, f=f) ## Perform adjusted test with `a.dm`

}
\seealso{
\code{\link{generic.distance.permutation.test}}, \code{\link{Tw2.test}},
         \code{\link{a.dist}}, \code{\link{dist.goodness.of.fit}}
}
